refactor: renamed files to lowercase_with_underscores
